Prefer the MACPORTS and HOMEBREW variables to the DETECT_XXX variables.
Its easier to read code that says "if MACPORTS" than code that says
"if DETECT_MACPORTS EQUAL 0". The former is obvious; the latter
requires you to remember that the variable contains the output of a
shell command and that "0" means "true" in this case even though it
means "false" everywhere else in computing.